- Changed wifi id from byte to int.

+++ ClientProject/ClientSource/DroneWifiDispatcher.cs
@@ -6,7 +6,7 @@ public partial class DroneWifiDispatcher : IServerSerializable
 {
     public void ClientEventRead(IReadMessage msg, float sendingTime)
     {
-        DroneId = msg.ReadByte();
+        DroneId = msg.ReadInt32();
         SendIdChannels();
     }
 }
